refactor: move latency indicator to navbar and cleanup trackpad ui
